Refactored reset password flow
