+++ src/components/ResetPasswordForm/index.js
@@ -3,7 +3,8 @@ import EnterEmail from './EnterEmail'
 import EnterNewPassword from './EnterNewPassword'
 
 export {
-  ResetPasswordForm,
   EnterEmail,
   EnterNewPassword,
 }
+
+export default ResetPasswordForm
