+++ src/components/index.js
@@ -3,7 +3,7 @@ import MFAConfirmForm from './MFAConfirmForm'
 import MFARecoveryForm from './MFARecoveryForm'
 import MFASetupForm from './MFASetupForm'
 import RegistrationForm from './RegistrationForm'
-import { ResetPasswordForm } from './ResetPassword'
+import ResetPasswordForm from './ResetPasswordForm'
 import UserProfileForm from './UserProfileForm'
 import UserProfileActions from './UserProfileActions'
 import VerifyEmailForm from './VerifyEmailForm'
